Move some DTLS read code into the read record layer
