+++ ssl/d1_lib.c
@@ -848,6 +848,7 @@ int DTLSv1_listen(SSL *ssl, BIO_ADDR *client)
         BIO_ADDR_clear(client);
 
     /* Buffer the record in the processed_rcds queue */
+    /* TODO(RECLAYER): This is nasty and reaches inside the record layer. FIXME */
     if (!dtls_buffer_listen_record(s, reclen, seq, align))
         return -1;
