+++ ssl/record/build.info
@@ -12,7 +12,7 @@ ENDIF
 
 SOURCE[../../libssl]=\
         rec_layer_s3.c rec_layer_d1.c ssl3_buffer.c  ssl3_record.c \
-        dtls1_bitmap.c ssl3_record_tls13.c
+        ssl3_record_tls13.c
 
 # For shared builds we need to include the sources needed in providers
 # (tls_pad.c) in libssl as well.
