+++ ssl/record/ssl3_record.c
@@ -833,430 +833,3 @@ int tls1_mac_old(SSL_CONNECTION *sc, SSL3_RECORD *rec, unsigned char *md,
     EVP_MD_CTX_free(hmac);
     return ret;
 }
-
-int dtls1_process_record(SSL_CONNECTION *s, DTLS1_BITMAP *bitmap)
-{
-    int i;
-    int enc_err;
-    SSL_SESSION *sess;
-    SSL3_RECORD *rr;
-    int imac_size;
-    size_t mac_size = 0;
-    unsigned char md[EVP_MAX_MD_SIZE];
-    size_t max_plain_length = SSL3_RT_MAX_PLAIN_LENGTH;
-    SSL_MAC_BUF macbuf = { NULL, 0 };
-    int ret = 0;
-    SSL *ssl = SSL_CONNECTION_GET_SSL(s);
-
-    rr = RECORD_LAYER_get_rrec(&s->rlayer);
-    sess = s->session;
-
-    /*
-     * At this point, s->rlayer.packet_length == SSL3_RT_HEADER_LNGTH + rr->length,
-     * and we have that many bytes in s->rlayer.packet
-     */
-    rr->input = &(s->rrlmethod->get0_packet(s->rrl)[DTLS1_RT_HEADER_LENGTH]);
-
-    /*
-     * ok, we can now read from 's->rlayer.packet' data into 'rr'. rr->input
-     * points at rr->length bytes, which need to be copied into rr->data by
-     * either the decryption or by the decompression. When the data is 'copied'
-     * into the rr->data buffer, rr->input will be pointed at the new buffer
-     */
-
-    /*
-     * We now have - encrypted [ MAC [ compressed [ plain ] ] ] rr->length
-     * bytes of encrypted compressed stuff.
-     */
-
-    /* check is not needed I believe */
-    if (rr->length > SSL3_RT_MAX_ENCRYPTED_LENGTH) {
-        SSLfatal(s, SSL_AD_RECORD_OVERFLOW, SSL_R_ENCRYPTED_LENGTH_TOO_LONG);
-        return 0;
-    }
-
-    /* decrypt in place in 'rr->input' */
-    rr->data = rr->input;
-    rr->orig_len = rr->length;
-
-    if (s->read_hash != NULL) {
-        const EVP_MD *tmpmd = EVP_MD_CTX_get0_md(s->read_hash);
-
-        if (tmpmd != NULL) {
-            imac_size = EVP_MD_get_size(tmpmd);
-            if (!ossl_assert(imac_size >= 0 && imac_size <= EVP_MAX_MD_SIZE)) {
-                    SSLfatal(s, SSL_AD_INTERNAL_ERROR, ERR_R_EVP_LIB);
-                    return 0;
-            }
-            mac_size = (size_t)imac_size;
-        }
-    }
-
-    if (SSL_READ_ETM(s) && s->read_hash) {
-        unsigned char *mac;
-
-        if (rr->orig_len < mac_size) {
-            SSLfatal(s, SSL_AD_DECODE_ERROR, SSL_R_LENGTH_TOO_SHORT);
-            return 0;
-        }
-        rr->length -= mac_size;
-        mac = rr->data + rr->length;
-        i = ssl->method->ssl3_enc->mac(s, rr, md, 0 /* not send */ );
-        if (i == 0 || CRYPTO_memcmp(md, mac, (size_t)mac_size) != 0) {
-            SSLfatal(s, SSL_AD_BAD_RECORD_MAC,
-                     SSL_R_DECRYPTION_FAILED_OR_BAD_RECORD_MAC);
-            return 0;
-        }
-        /*
-         * We've handled the mac now - there is no MAC inside the encrypted
-         * record
-         */
-        mac_size = 0;
-    }
-
-    /*
-     * Set a mark around the packet decryption attempt.  This is DTLS, so
-     * bad packets are just ignored, and we don't want to leave stray
-     * errors in the queue from processing bogus junk that we ignored.
-     */
-    ERR_set_mark();
-    enc_err = ssl->method->ssl3_enc->enc(s, rr, 1, 0, &macbuf, mac_size);
-
-    /*-
-     * enc_err is:
-     *    0: if the record is publicly invalid, or an internal error, or AEAD
-     *       decryption failed, or ETM decryption failed.
-     *    1: Success or MTE decryption failed (MAC will be randomised)
-     */
-    if (enc_err == 0) {
-        ERR_pop_to_mark();
-        if (ossl_statem_in_error(s)) {
-            /* SSLfatal() got called */
-            goto end;
-        }
-        /* For DTLS we simply ignore bad packets. */
-        rr->length = 0;
-        s->rrlmethod->reset_packet_length(s->rrl);
-        goto end;
-    }
-    ERR_clear_last_mark();
-    OSSL_TRACE_BEGIN(TLS) {
-        BIO_printf(trc_out, "dec %zd\n", rr->length);
-        BIO_dump_indent(trc_out, rr->data, rr->length, 4);
-    } OSSL_TRACE_END(TLS);
-
-    /* r->length is now the compressed data plus mac */
-    if ((sess != NULL)
-            && !SSL_READ_ETM(s)
-            && (s->enc_read_ctx != NULL)
-            && (EVP_MD_CTX_get0_md(s->read_hash) != NULL)) {
-        /* s->read_hash != NULL => mac_size != -1 */
-
-        i = ssl->method->ssl3_enc->mac(s, rr, md, 0 /* not send */ );
-        if (i == 0 || macbuf.mac == NULL
-            || CRYPTO_memcmp(md, macbuf.mac, mac_size) != 0)
-            enc_err = 0;
-        if (rr->length > SSL3_RT_MAX_COMPRESSED_LENGTH + mac_size)
-            enc_err = 0;
-    }
-
-    if (enc_err == 0) {
-        /* decryption failed, silently discard message */
-        rr->length = 0;
-        s->rrlmethod->reset_packet_length(s->rrl);
-        goto end;
-    }
-
-    /* r->length is now just compressed */
-    if (s->expand != NULL) {
-        if (rr->length > SSL3_RT_MAX_COMPRESSED_LENGTH) {
-            SSLfatal(s, SSL_AD_RECORD_OVERFLOW,
-                     SSL_R_COMPRESSED_LENGTH_TOO_LONG);
-            goto end;
-        }
-        if (!ssl3_do_uncompress(s, rr)) {
-            SSLfatal(s, SSL_AD_DECOMPRESSION_FAILURE, SSL_R_BAD_DECOMPRESSION);
-            goto end;
-        }
-    }
-
-    /* use current Max Fragment Length setting if applicable */
-    if (s->session != NULL && USE_MAX_FRAGMENT_LENGTH_EXT(s->session))
-        max_plain_length = GET_MAX_FRAGMENT_LENGTH(s->session);
-
-    /* send overflow if the plaintext is too long now it has passed MAC */
-    if (rr->length > max_plain_length) {
-        SSLfatal(s, SSL_AD_RECORD_OVERFLOW, SSL_R_DATA_LENGTH_TOO_LONG);
-        goto end;
-    }
-
-    rr->off = 0;
-    /*-
-     * So at this point the following is true
-     * ssl->s3.rrec.type   is the type of record
-     * ssl->s3.rrec.length == number of bytes in record
-     * ssl->s3.rrec.off    == offset to first valid byte
-     * ssl->s3.rrec.data   == where to take bytes from, increment
-     *                        after use :-).
-     */
-
-    /* we have pulled in a full packet so zero things */
-    s->rrlmethod->reset_packet_length(s->rrl);
-
-    /* Mark receipt of record. */
-    dtls1_record_bitmap_update(s, bitmap);
-
-    ret = 1;
- end:
-    if (macbuf.alloced)
-        OPENSSL_free(macbuf.mac);
-    return ret;
-}
-
-/*
- * Retrieve a buffered record that belongs to the current epoch, i.e. processed
- */
-#define dtls1_get_processed_record(s) \
-                   dtls1_retrieve_buffered_record((s), \
-                   &(DTLS_RECORD_LAYER_get_processed_rcds(&s->rlayer)))
-
-/*-
- * Call this to get a new input record.
- * It will return <= 0 if more data is needed, normally due to an error
- * or non-blocking IO.
- * When it finishes, one packet has been decoded and can be found in
- * ssl->s3.rrec.type    - is the type of record
- * ssl->s3.rrec.data    - data
- * ssl->s3.rrec.length  - number of bytes
- */
-/* used only by dtls1_read_bytes */
-int dtls1_get_record(SSL_CONNECTION *s)
-{
-    int ssl_major, ssl_minor;
-    int rret;
-    size_t more, n;
-    SSL3_RECORD *rr;
-    unsigned char *p = NULL;
-    unsigned short version;
-    DTLS1_BITMAP *bitmap;
-    unsigned int is_next_epoch;
-    SSL *ssl = SSL_CONNECTION_GET_SSL(s);
-
-    rr = RECORD_LAYER_get_rrec(&s->rlayer);
-
- again:
-    /*
-     * The epoch may have changed.  If so, process all the pending records.
-     * This is a non-blocking operation.
-     */
-    if (!dtls1_process_buffered_records(s)) {
-        /* SSLfatal() already called */
-        return -1;
-    }
-
-    /* if we're renegotiating, then there may be buffered records */
-    if (dtls1_get_processed_record(s))
-        return 1;
-
-    /* get something from the wire */
-
-    /* check if we have the header */
-    if ((RECORD_LAYER_get_rstate(&s->rlayer) != SSL_ST_READ_BODY) ||
-        (s->rrlmethod->get_packet_length(s->rrl) < DTLS1_RT_HEADER_LENGTH)) {
-        rret = HANDLE_RLAYER_RETURN(s,
-            s->rrlmethod->read_n(s->rrl, DTLS1_RT_HEADER_LENGTH,
-                                 SSL3_BUFFER_get_len(s->rrlmethod->get0_rbuf(s->rrl)), 0, 1, &n));
-        /* read timeout is handled by dtls1_read_bytes */
-        if (rret <= 0) {
-            /* SSLfatal() already called if appropriate */
-            return rret;         /* error or non-blocking */
-        }
-
-        /* this packet contained a partial record, dump it */
-        if (s->rrlmethod->get_packet_length(s->rrl) != DTLS1_RT_HEADER_LENGTH) {
-            s->rrlmethod->reset_packet_length(s->rrl);
-            goto again;
-        }
-
-        RECORD_LAYER_set_rstate(&s->rlayer, SSL_ST_READ_BODY);
-
-        p = s->rrlmethod->get0_packet(s->rrl);
-
-        if (s->msg_callback)
-            s->msg_callback(0, 0, SSL3_RT_HEADER, p, DTLS1_RT_HEADER_LENGTH,
-                            ssl, s->msg_callback_arg);
-
-        /* Pull apart the header into the DTLS1_RECORD */
-        rr->type = *(p++);
-        ssl_major = *(p++);
-        ssl_minor = *(p++);
-        version = (ssl_major << 8) | ssl_minor;
-
-        /* sequence number is 64 bits, with top 2 bytes = epoch */
-        n2s(p, rr->epoch);
-
-        memcpy(&(RECORD_LAYER_get_read_sequence(&s->rlayer)[2]), p, 6);
-        p += 6;
-
-        n2s(p, rr->length);
-        rr->read = 0;
-
-        /*
-         * Lets check the version. We tolerate alerts that don't have the exact
-         * version number (e.g. because of protocol version errors)
-         */
-        if (!s->first_packet && rr->type != SSL3_RT_ALERT) {
-            if (version != s->version) {
-                /* unexpected version, silently discard */
-                rr->length = 0;
-                rr->read = 1;
-                s->rrlmethod->reset_packet_length(s->rrl);
-                goto again;
-            }
-        }
-
-        if ((version & 0xff00) != (s->version & 0xff00)) {
-            /* wrong version, silently discard record */
-            rr->length = 0;
-            rr->read = 1;
-            s->rrlmethod->reset_packet_length(s->rrl);
-            goto again;
-        }
-
-        if (rr->length > SSL3_RT_MAX_ENCRYPTED_LENGTH) {
-            /* record too long, silently discard it */
-            rr->length = 0;
-            rr->read = 1;
-            s->rrlmethod->reset_packet_length(s->rrl);
-            goto again;
-        }
-
-        /* If received packet overflows own-client Max Fragment Length setting */
-        if (s->session != NULL && USE_MAX_FRAGMENT_LENGTH_EXT(s->session)
-                && rr->length > GET_MAX_FRAGMENT_LENGTH(s->session) + SSL3_RT_MAX_ENCRYPTED_OVERHEAD) {
-            /* record too long, silently discard it */
-            rr->length = 0;
-            rr->read = 1;
-            s->rrlmethod->reset_packet_length(s->rrl);
-            goto again;
-        }
-
-        /* now s->rlayer.rstate == SSL_ST_READ_BODY */
-    }
-
-    /* s->rlayer.rstate == SSL_ST_READ_BODY, get and decode the data */
-
-    if (rr->length >
-        s->rrlmethod->get_packet_length(s->rrl) - DTLS1_RT_HEADER_LENGTH) {
-        /* now s->rlayer.packet_length == DTLS1_RT_HEADER_LENGTH */
-        more = rr->length;
-        rret = HANDLE_RLAYER_RETURN(s,
-            s->rrlmethod->read_n(s->rrl, more, more, 1, 1, &n));
-        /* this packet contained a partial record, dump it */
-        if (rret <= 0 || n != more) {
-            if (ossl_statem_in_error(s)) {
-                /* read_n() called SSLfatal() */
-                return -1;
-            }
-            rr->length = 0;
-            rr->read = 1;
-            s->rrlmethod->reset_packet_length(s->rrl);
-            goto again;
-        }
-
-        /*
-         * now n == rr->length, and s->rlayer.packet_length ==
-         * DTLS1_RT_HEADER_LENGTH + rr->length
-         */
-    }
-    /* set state for later operations */
-    RECORD_LAYER_set_rstate(&s->rlayer, SSL_ST_READ_HEADER);
-
-    /* match epochs.  NULL means the packet is dropped on the floor */
-    bitmap = dtls1_get_bitmap(s, rr, &is_next_epoch);
-    if (bitmap == NULL) {
-        rr->length = 0;
-        s->rrlmethod->reset_packet_length(s->rrl); /* dump this record */
-        goto again;             /* get another record */
-    }
-#ifndef OPENSSL_NO_SCTP
-    /* Only do replay check if no SCTP bio */
-    if (!BIO_dgram_is_sctp(SSL_get_rbio(ssl))) {
-#endif
-        /* Check whether this is a repeat, or aged record. */
-        if (!dtls1_record_replay_check(s, bitmap)) {
-            rr->length = 0;
-            rr->read = 1;
-            s->rrlmethod->reset_packet_length(s->rrl); /* dump this record */
-            goto again;         /* get another record */
-        }
-#ifndef OPENSSL_NO_SCTP
-    }
-#endif
-
-    /* just read a 0 length packet */
-    if (rr->length == 0) {
-        rr->read = 1;
-        goto again;
-    }
-
-    /*
-     * If this record is from the next epoch (either HM or ALERT), and a
-     * handshake is currently in progress, buffer it since it cannot be
-     * processed at this time.
-     */
-    if (is_next_epoch) {
-        if ((SSL_in_init(ssl) || ossl_statem_get_in_handshake(s))) {
-            if (dtls1_buffer_record (s,
-                    &(DTLS_RECORD_LAYER_get_unprocessed_rcds(&s->rlayer)),
-                    rr->seq_num) < 0) {
-                /* SSLfatal() already called */
-                return -1;
-            }
-        }
-        rr->length = 0;
-        rr->read = 1;
-        s->rrlmethod->reset_packet_length(s->rrl);
-        goto again;
-    }
-
-    if (!dtls1_process_record(s, bitmap)) {
-        if (ossl_statem_in_error(s)) {
-            /* dtls1_process_record() called SSLfatal */
-            return -1;
-        }
-        rr->length = 0;
-        rr->read = 1;
-        s->rrlmethod->reset_packet_length(s->rrl); /* dump this record */
-        goto again;             /* get another record */
-    }
-
-    return 1;
-
-}
-
-int dtls_buffer_listen_record(SSL_CONNECTION *s, size_t len, unsigned char *seq,
-                              size_t off)
-{
-    SSL3_RECORD *rr;
-
-    rr = RECORD_LAYER_get_rrec(&s->rlayer);
-    memset(rr, 0, sizeof(SSL3_RECORD));
-
-    rr->length = len;
-    rr->type = SSL3_RT_HANDSHAKE;
-    memcpy(rr->seq_num, seq, sizeof(rr->seq_num));
-    rr->off = off;
-
-    s->rrlmethod->set0_packet(s->rrl, s->rrlmethod->get0_rbuf(s->rrl)->buf,
-                              DTLS1_RT_HEADER_LENGTH + len);
-    rr->data = s->rrlmethod->get0_packet(s->rrl) + DTLS1_RT_HEADER_LENGTH;
-
-    if (dtls1_buffer_record(s, &(s->rlayer.d->processed_rcds),
-                            SSL3_RECORD_get_seq_num(s->rlayer.rrec)) <= 0) {
-        /* SSLfatal() already called */
-        return 0;
-    }
-
-    return 1;
-}
